+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
 # Changelog
 
-## v?.?.? rpetit3/sizemeup "" 2024/??/??
+## v1.0.0 rpetit3/sizemeup "Tape Measure" 2024/09/29
+
+- `sizemeup-build` to build the data for the `sizemeup` package
+- `sizemeup` to output known genome size for a species